araml/requirements.txt
torch>=2.0.0
transformers>=4.38.0
datasets>=2.18.0
faiss-cpu>=1.7.4
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
tqdm>=4.66.0
pyyaml>=6.0.1
higher>=0.2.1
wandb>=0.16.0
jupyter>=1.0.0
matplotlib>=3.8.0
pandas>=2.1.0
